commit after generalisation, pass test
