+++ src/AppBundle/Calculator/Mk2Calculator.php
@@ -4,27 +4,13 @@ namespace AppBundle\Calculator;
 
 use AppBundle\Model\Change;
 
-class Mk2Calculator implements CalculatorInterface {
+class Mk2Calculator extends Calculator implements CalculatorInterface {
+
+  public function __construct() {
+    parent::__construct( ['coin1']);
+  }
 
   public function getSupportedModel(): string {
     return 'mk2';
-  }
-
-  /**
-   * @param int $amount The amount of money to turn into change
-   *
-   * @return Change|null The change, or null if the operation is impossible
-   */
-  public function getChange(int $amount): ? Change {
-    $params = Change::params;
-    unset($params['coin1']);
-
-    $reste = $amount;
-    $change = new Change;
-    foreach( $params as $changeAttribut => $denominateur ) {
-      $change->$changeAttribut = (int) floor( $reste / $denominateur );
-      $reste = $reste % $denominateur;
-    }
-    return $change->getAmount() == $amount ? $change : null;
   }
 }
